Import clinical information for a patient

Import clinical information for a patient

import_clinical(clinicalData = NA, patientID = NA,
  extract_column = "Blast")

Arguments

clinicalData
data frame containing the clinical information for a patient. This needs to include a column `SampleName` containing unique names for all the RNA-Seq samples available in the cohort; a `PID` columns for patient ID and a `Time` column that should reflect the time of sample collection.

patientID
a character vector specifying the patient id for which clinical data should be imported.

extract_column
vector specifying the column to be extracted. Currently only numeric variables with values between 0 and 1 are allowed to be consistent with plotting together with clonality and VAF estimates.

Examples

clinicalData <- data.frame(SampleName = c("D1.Screen.Rel","D1.Cyc1.Rel",
                                          "D1.Cyc2.Rel","D1.Cyc3.Rel"),
                           PID = "D1",
                           Time = c("Screen", "Cyc1","Cyc2","Cyc3"),
                           Outcome = "Rel",
                           AgeDiagnosis = 65,
                           Sex = "F",
                           Blast = c(80,5,7,40)/100)

import_blast <- import_clinical(clinicalData = clinicalData,
extract_column = "Blast",
patientID = "D1")